Add citation for the inverse ladder to the readme.
Pieter was inspired by @peterdettman's faster inverse and sqrt ladders in libsecp256k1 to reinvent the technique of this 1988 paper.  The work that inspired Dettman's effort's cited the 1988 paper, so we went full loop.

I ran across the paper while looking at other work on fast inverses and thought to check if it could be applied to make minisketch faster... apparently not! :)